Represent large sequence literals using the recently added ArrayLiteral support (#1317)
The recently added ArrayLiteral support offers a path towards a more efficient encoding for sequence literals (compared to a push-based description), since it directly encodes the value at each index of the literal. This PR updates the seq! macro to use a view of an equivalent array literal for sequence literals with more than one element. It also updates the interpreter to match this behavior in its specialized internal handling of sequences and arrays.